Remove old core, update setup.py
