--- .gitattributes
@@ -1,1 +1,0 @@
-tensorhive/static/* linguist-vendored
