--- MANIFEST.in
@@ -1,2 +1,0 @@
-recursive-include tensorhive/static *
-recursive-include tensorhive/scripts *
