+++ tensorhive/core/managers/ServiceManager.py
@@ -1,8 +1,8 @@
 # types
-from tensorhive.core_anew.managers.InfrastructureManager import InfrastructureManager
-from tensorhive.core_anew.managers.SSHConnectionManager import SSHConnectionManager
+from tensorhive.core.managers.InfrastructureManager import InfrastructureManager
+from tensorhive.core.managers.SSHConnectionManager import SSHConnectionManager
 
-from tensorhive.core_anew.services.Service import Service
+from tensorhive.core.services.Service import Service
 from typing import List
 
 class ServiceManager():
