+++ tensorhive/core/managers/TensorHiveManager.py
@@ -1,16 +1,16 @@
 from threading import Thread
-from tensorhive.core_anew.managers.ThreadedManager import ThreadedManager
-from tensorhive.core_anew.managers.InfrastructureManager import InfrastructureManager
-from tensorhive.core_anew.managers.SSHConnectionManager import SSHConnectionManager
-from tensorhive.core_anew.managers.ServiceManager import ServiceManager
+from tensorhive.core.managers.ThreadedManager import ThreadedManager
+from tensorhive.core.managers.InfrastructureManager import InfrastructureManager
+from tensorhive.core.managers.SSHConnectionManager import SSHConnectionManager
+from tensorhive.core.managers.ServiceManager import ServiceManager
 
 
-from tensorhive.core_anew.services.Service import Service
+from tensorhive.core.services.Service import Service
 from typing import List, Dict
 
 # FIXME Remove colorama dependency from setup.py
 from colorama import Fore, Back, Style
-from tensorhive.core_anew.utils.decorators.override import override
+from tensorhive.core.utils.decorators.override import override
 from tensorhive.config import CONFIG, SSH_CONFIG, API_CONFIG
 from tensorhive.api.APIServer import APIServer
 
@@ -24,7 +24,7 @@ class TensorHiveManager(ThreadedManager):
         self.infrastructure_manager = InfrastructureManager()
 
         # FIXME hardcoded SSHConnector
-        print(f'{Back.RED}WARNING! You need to replace hostname and username in tensorhive/core_anew/managers/TensorHiveManager.py{Style.RESET_ALL}')
+        print(f'{Back.RED}WARNING! You need to replace hostname and username in tensorhive/core/managers/TensorHiveManager.py{Style.RESET_ALL}')
 
         self.connection_manager = SSHConnectionManager(
             SSH_CONFIG.AVAILABLE_NODES)
