--- tensorhive/core/monitoring_handlers/PrintingHandler.py
@@ -1,7 +1,0 @@
-from tensorhive.core.monitoring import MonitoringHandler
-
-
-class PrintingHandler(MonitoringHandler):
-    def handle_monitoring(self, infrastructure):
-        print(str(infrastructure))
-
