+++ tensorhive/core/utils/StoppableThread.py
